requirements.txt
Flask
gunicorn
langchain
google-cloud-aiplatform